fix(pagos): la multa de penalización y la referencia de pago del alta
Dos bugs REALES en código desplegado, encontrados al poner al día los tests que
llevaban rojos desde el 12 de agosto. Las pruebas tenían razón.

1. La multa por penalización nunca se pudo emitir. En debt-scheduler el
   parámetro penalty_weeks se usa antes en `p.price_usd * $1`, lo que tipa el
   parámetro como numeric para toda la sentencia, y make_interval solo acepta
   enteros: "function make_interval(weeks => numeric) does not exist". El tick
   COMPLETO abortaba ahí, así que el chofer quedaba penalizado pero sin multa y
   sin los registros de auditoría de esa pasada. Casteo explícito a ::int.

2. El método y la referencia de un alta hecha por el admin no se veían en
   ninguna parte. El listado y el detalle de facturas los leían solo desde el
   recibo, pero cuando el admin inscribe directo no hay recibo: se estampan en
   las columnas de la factura. Ese escritor quedó huérfano al mudar la lectura.
   COALESCE(recibo, factura) en ambas consultas + hasProof mira el proof_url
   heredado.

Tests: 19/30 -> 29/30. Todos los arreglados describían diseños cambiados a
propósito (factura por concepto, external-payment eliminado, cargos generados al
crear el recibo).

El que queda rojo ROTA entre dos pruebas del motor y siempre por un cargo de
más: lo emite el scheduler de producción sobre la misma base mientras los tests
corren. No se arregla aflojando aserciones de dinero.

+++ src/plugins/debt-scheduler.ts
@@ -265,7 +265,7 @@ export async function runDebtEngineTick(db: pg.Pool): Promise<DebtTickResult> {
            INSERT INTO subscription_payments
              (driver_subscription_id, invoice_id, period_start, period_end,
               amount_usd, status, charge_kind)
-           SELECT e.sub_id, ni.invoice_id, now(), now() + make_interval(weeks => $1),
+           SELECT e.sub_id, ni.invoice_id, now(), now() + make_interval(weeks => $1::int),
                   e.amount, 'pending', 'penalty'
            FROM eligible e JOIN new_invoices ni ON ni.driver_id = e.driver_id
            RETURNING id, driver_subscription_id
